add autocomplete to search control

+++ app/src/main/java/com/safrzone/safrzone/services/MapBoxService.java
@@ -41,7 +41,7 @@ public interface MapBoxService {
         public String placeName;
 
         @SerializedName("center")
-        public List<Double> center;
+        public List<Float> center;
     }
 }
